De-duplicate command metadata storage classes
